Cleanup fallback image logic
